ajusta arquivo sql

+++ database.sql
@@ -8,7 +8,7 @@ drop type if exists usuario_type;
 drop type if exists categoria_type;
 
 CREATE TABLE usuario(
-   id INT GENERATED ALWAYS AS IDENTITY,
+   id BIGINT  GENERATED ALWAYS AS IDENTITY,
    nome VARCHAR(255) NOT NULL,
    email VARCHAR(255) NOT NULL,
    senha VARCHAR(255) NOT NULL,
@@ -17,7 +17,7 @@ CREATE TABLE usuario(
 );
 
 CREATE TABLE localizacao(
-   id INT GENERATED ALWAYS AS IDENTITY,
+   id BIGINT GENERATED ALWAYS AS IDENTITY,
    pais VARCHAR(255) NOT NULL,
    estado VARCHAR(255) NOT NULL,
    cidade VARCHAR(255) NOT NULL,
@@ -25,7 +25,7 @@ CREATE TABLE localizacao(
 );
 
 CREATE TABLE atracao(
-   id INT GENERATED ALWAYS AS IDENTITY,
+   id BIGINT GENERATED ALWAYS AS IDENTITY,
    nome VARCHAR(255) NOT NULL,
    descricao VARCHAR NOT NULL,
    foto VARCHAR NOT NULL,
@@ -38,7 +38,7 @@ CREATE TABLE atracao(
 );
 
 CREATE TABLE avaliacao_atracao(
-   id INT GENERATED ALWAYS AS IDENTITY,
+   id BIGINT  GENERATED ALWAYS AS IDENTITY,
    positivo BOOLEAN NOT NULL,
    negativo BOOLEAN NOT NULL,
    usuario_id INT,
@@ -49,7 +49,7 @@ CREATE TABLE avaliacao_atracao(
 );
 
 CREATE TABLE recomendacao(
-   id INT GENERATED ALWAYS AS IDENTITY,
+   id BIGINT GENERATED ALWAYS AS IDENTITY,
    conteudo VARCHAR NOT NULL,
    usuario_id INT,
    atracao_id INT,
@@ -59,7 +59,7 @@ CREATE TABLE recomendacao(
 );
 
 CREATE TABLE avaliacao_recomendacao(
-   id INT GENERATED ALWAYS AS IDENTITY,
+   id BIGINT GENERATED ALWAYS AS IDENTITY,
    positivo BOOLEAN NOT NULL,
    negativo BOOLEAN NOT NULL,
    usuario_id INT,
